+++ Forms/CommandLineForm.vb
@@ -39,12 +39,12 @@ Public Class CommandLineForm
         cbGoTo.Select()
 
         cms.Add("Execute Command Line", Sub() params.Execute(), p.SourceFile <> "").SetImage(Symbol.fa_terminal)
-        cms.Add("Copy Command Line", Sub() Clipboard.SetText(params.GetCommandLine(True, True)))
+        cms.Add("Copy Command Line", Sub() Clipboard.SetText(params.GetCommandLine(True, True))).SetImage(Symbol.Copy)
         cms.Add("Show Command Line...", Sub() g.ShowCommandLinePreview("Command Line", params.GetCommandLine(True, True)))
-        cms.Add("Import Command Line...", Sub() If MsgQuestion("Import command line from clipboard?", Clipboard.GetText) = DialogResult.OK Then BasicVideoEncoder.ImportCommandLine(Clipboard.GetText, params))
+        cms.Add("Import Command Line...", Sub() If MsgQuestion("Import command line from clipboard?", Clipboard.GetText) = DialogResult.OK Then BasicVideoEncoder.ImportCommandLine(Clipboard.GetText, params)).SetImage(Symbol.Download)
 
-        cms.Add("Help", AddressOf ShowHelp).SetImage(Symbol.Help)
-        cms.Add(params.GetPackage.Name + " Help", Sub() g.StartProcess(params.GetPackage.GetHelpPath))
+        cms.Add("Help about this dialog", AddressOf ShowHelp).SetImage(Symbol.Help)
+        cms.Add("Help about " + params.GetPackage.Name, Sub() params.GetPackage.ShowHelp()).SetImage(Symbol.Help)
     End Sub
 
     Sub SelectLastPage()